roachtest: embed helper files instead of using c.Put
There are a few roachtests that rely on config or helper
files that must be copied onto the crdb nodes for the test
to function.

A previous approach to this was to use c.Put() to copy
the file from the test runner to the node. However, this
requires hardcoding the relative path of the file from the
roachtest runner.

By embedding these files instead, the dependency switches
to where the helper file lives relative to the test that
needs it, something that is much less likely to change.

Release note: none
Fixes: none
Epic: none
